Migration of metal3 helm charts from rancher-sandbox (#6)
* Migrated the metal3 helm charts from the rancher-sandbox org

* started migrating

Began the migration from metal3-deploy charts into our own repo. Mostly everything works, however, ran into issues with external-dns and now with something in the chart. Committing this as a placeholder as I make changes

* finishing up

Everything is where it should be now. Just need some finishing tweaks. For some reason on OSX, helm is bugging with the plugins when running "make html"

Gonna push this so I can open it on a linux server and run the command there

* ran "make html"

* updated read me, not sure if correct

* accidentally added this, removing in this commit

* delete ex-dns

* comment changes

- changed external-dns from local to upstream
- added better descriptions to each dependency/chart
- added warning to package/akri/gen/overlay/app-readme.md
- tried to add app-readme.md to metal3 but when i ran makechart it wouldn't go to the generated charts file

also, not sure what to put in the readme yet for metal3 as i've yet to try it

* moved app-readme

---------

Co-authored-by: Jason Dobies <jason_dobies@yahoo.com>