metal3: fix config for disabled provisioning network
Currently the defaults assume a provisioning network is enabled, and
it's not possible to disable because some containers crashloop
waiting for PROVISIONING_IP to be ready.

To resolve this we can optionally set the PROVISIONING_IP to the
hostIP, which allows a configuration like this to run without any
dedicated provisioning network:

  global:
    enable_dnsmasq: false
    enable_pxe_boot: false
    provisioningInterface: ""
    provisioningIP: ""

Note this will only work for virtualmedia deployments since PXE is
disabled in this mode.